requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
PyJWT==2.9.0
eventlet==0.36.1
pandas==2.3.3
